agents/documentationAgent.py
"""
Autonomous Documentation Agent
Generates industry-specific compliance docs on demand
"""
import anthropic
import os

class DocumentationAgent:
    def __init__(self):
        self.client = anthropic.Anthropic(
            api_key=os.environ.get("ANTHROPIC_API_KEY")
        )
    
    def generate_industry_guide(self, industry, framework):
        """
        Generate compliance guide for specific industry
        Example: generate_industry_guide("healthcare", "HIPAA")
        """
        
        prompt = f"""Generate a comprehensive compliance guide for {industry} 
        using the {framework} framework, specifically focused on Midnight blockchain 
        integration. Include:
        
        1. Regulatory requirements
        2. Midnight's privacy features that address compliance
        3. Implementation steps
        4. Code examples
        5. Audit considerations
        
        Make it actionable and specific to blockchain implementation."""
        
        message = self.client.messages.create(
            model="claude-sonnet-4-20250514",
            max_tokens=4000,
            messages=[
                {"role": "user", "content": prompt}
            ]
        )
        
        return message.content[0].text
    
    def generate_onboarding_workflow(self, company_profile):
        """
        Generate custom onboarding workflow based on company needs
        """
        pass
    
    def update_documentation(self, category):
        """
        Automatically update docs with latest Midnight features
        """
        pass

agent = DocumentationAgent()
