requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
blockfrost-python==0.6.0
requests==2.32.5
beautifulsoup4==4.12.3
python-dotenv==1.0.0
jsonschema==4.25.1
